+++ src/frontend/step_definitions/01-test_steps.ts
@@ -1,7 +1,0 @@
-import {createBdd} from "playwright-bdd"
-const {Given, When, Then} = createBdd();
-
-
-Given(/^test step$/, async function ({page}) {
-
-});